CHORE: refactored tdf spectrum reader
